requirements.txt
google-auth
earthengine-api
google-generativeai
python-dotenv
geemap
fastapi
uvicorn